Big rewrite to use a different MQTT library which has retain and seems better maintained. Also support setting individual params on specific topics instead of only allowing a single JSON blob.

+++ Aircon_Remote/DL_Aircon.h
@@ -7,7 +7,7 @@ typedef struct {
   bool on;
   bool oscillate;
   bool timer;
-  int timer_value;
+  float timer_value;
   int speed;
   int wind;
 } dl_aircon_msg_t;
